integer force/mass in Bullet class

+++ classes.py
@@ -101,7 +101,7 @@ class Bullet(object):
         self.pos = Vector2(x, y)
         self.vel = Vector2(0, 0)
         self.hitbox = pygame.Rect(x - width / 2, y - height / 2, width, height)
-        acc = force / mass
+        acc = int(force / mass)
         self.acc = Vector2(0, acc)
 
         self.game = game
